WpRole - Autorisations & modif migrations & modif show wp users

See merge request pfe-2023/ubo/back!30

+++ app/Http/Controllers/WpRoleController.php
@@ -39,6 +39,8 @@ class WpRoleController extends Controller
      */
     public function store(StoreWpRoleRequest $request): JsonResponse
     {
+       $this->authorize('createWpRole',  WpRole::class);
+
         return $this->wpRoleService->storeWpRole($request->all());
     }
     
@@ -61,6 +63,8 @@ class WpRoleController extends Controller
      */
     public function update(UpdateWpRoleRequest $request, WpRole $wpRole): JsonResponse
     {
+        $this->authorize('updateWpRole',  WpRole::class);
+
         return $this->wpRoleService->updateWpRole($request->all(), $wpRole);
     }
         
@@ -72,6 +76,8 @@ class WpRoleController extends Controller
      */
     public function destroy(WpRole $wpRole): JsonResponse
     {
+        $this->authorize('delete',  WpRole::class);
+
         return $this->wpRoleService->deleteWpRole($wpRole);
     }
 }
